Point system v4 - "gainPoints" metoden virker nu. Returner en point værdi

+++ zuul-framework/src/worldofzuul/Rewards.java
@@ -28,6 +28,7 @@ public class Rewards extends InteractiveObjects {
     //Method for gaining points
     public int gainPoints() {
 
+
         int pointsModifier;
 
         if (itemNeed == itemType)
@@ -41,7 +42,6 @@ public class Rewards extends InteractiveObjects {
 
         gainPoints = initialPoints * pointsModifier;
 
-
-    return gainPoints();
+    return (int) gainPoints;
     }
 }
